Create AWS API signing enpoint.
Provision server with AWS bucket credentials and provide this data to Django.
Part of this data is passed to the client settings so we can allow users to
upload files directly to S3 without having our server act as a middleman. AWS
requests from the client must be signed with our AWS secret key so we provide
and endpoint which will be used to validate our user is allowed to upload data
to our bucket.